rename files to remove redunant pybert_*
