+++ pybert/__init__.py
@@ -17,26 +17,26 @@ The application source is divided among several files, as follows:
                       - the definitions of those dependent variables, which are handled
                         automatically by the Traits/UI machinery.
 
-    pybert_view.py  - The V in MVC, it contains the main window layout definition, as
+    view.py  - The V in MVC, it contains the main window layout definition, as
                       well as the definitions of user invoked actions
                       (i.e.- buttons).
 
-    pybert_cntrl.py - The C in MVC, it contains the definitions for those dependent
+    control.py - The C in MVC, it contains the definitions for those dependent
                       variables, which are updated not automatically by
                       the Traits/UI machinery, but rather by explicit
                       user action (i.e. - button clicks).
 
-    pybert_help.py  - Contents for the "Help" tab of the GUI.
+    help.py  - Contents for the "Help" tab of the GUI.
 
-    pybert_plot.py  - Contains all plot definitions.
+    plot.py  - Contains all plot definitions.
 
-    pybert_util.py  - Contains general purpose utility functionality.
+    utility.py  - Contains general purpose utility functionality.
 
-    pybert_cfg.py   - Defines the data structure for storing PyBERT
+    configuration.py   - Defines the data structure for storing PyBERT
                       configurations, so they may be saved and later
                       restored.
 
-    pybert_data.py  - Defines the data structure for storing PyBERT
+    results.py  - Defines the data structure for storing PyBERT
                       simulation results, so they may be saved and later
                       recalled as reference waveforms for comparison.
