+++ pybert/control.py
@@ -41,7 +41,7 @@ from scipy.signal.windows import hann
 from pyibisami.ami_model import AMIModel, AMIModelInitializer
 
 from pybert.dfe import DFE
-from pybert.pybert_util import (
+from pybert.utility import (
     calc_eye,
     calc_jitter,
     find_crossings,
